requirements.txt
flask>=3.0.3
ollama>=0.2.1
opencv-python>=4.10.0.82
mediapipe>=0.10.14
numpy>=1.26.4,<2.0
crewai==0.5.0
sentence-transformers>=3.0.1
chromadb>=0.5.23
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-proto==1.18.0
opentelemetry-semantic-conventions>=0.37b0
ultralytics>=8.2.52
face_recognition>=1.3.0
pyaudio>=0.2.14
pydub>=0.25.1
keyring>=25.2.1
google-api-python-client>=2.135.0
msal>=1.29.0
googlemaps>=4.10.0
yfinance>=0.2.40
alpha-vantage>=3.0.0
finnhub-python>=2.4.20
pyicloud>=1.0.0
caldav>=1.3.9
requests>=2.32.3
torch>=2.3.1
pyttsx3>=2.90
paginate==0.5.6